reorganized metrics
