+++ pkg/controllers/state/pod.go
@@ -16,16 +16,12 @@ package state
 
 import (
 	"context"
-	"fmt"
-	"strings"
 	"sync"
 	"time"
 
-	"github.com/aws/karpenter/pkg/apis/provisioning/v1alpha5"
 	"github.com/prometheus/client_golang/prometheus"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/types"
 	"knative.dev/pkg/logging"
 	controllerruntime "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -83,56 +79,4 @@ func (c *PodController) Register(ctx context.Context, m manager.Manager) error {
 		Named(podControllerName).
 		For(&v1.Pod{}).
 		Complete(c)
-}
-
-func (c *PodController) record(ctx context.Context, pod *v1.Pod) {
-	logging.FromContext(ctx).Infof("Recording object: %s", client.ObjectKeyFromObject(pod).String())
-	labels := c.labels(ctx, pod)
-	podGaugeVec.With(labels).Set(float64(1))
-	c.labelsMap.Store(client.ObjectKeyFromObject(pod), labels)
-}
-
-// labels creates the labels using the current state of the pod
-func (c *PodController) labels(ctx context.Context, pod *v1.Pod) prometheus.Labels {
-	metricLabels := prometheus.Labels{}
-	metricLabels[podName] = pod.GetName()
-	metricLabels[podNameSpace] = pod.GetNamespace()
-	// Selflink has been deprecated after v.1.20
-	// Manually generate the selflink for the first owner reference
-	// Currently we do not support multiple owner references
-	selflink := ""
-	if len(pod.GetOwnerReferences()) > 0 {
-		ownerreference := pod.GetOwnerReferences()[0]
-		selflink = fmt.Sprintf("/apis/%s/namespaces/%s/%ss/%s", ownerreference.APIVersion, pod.Namespace, strings.ToLower(ownerreference.Kind), ownerreference.Name)
-	}
-	metricLabels[ownerSelfLink] = selflink
-	metricLabels[podHostName] = pod.Spec.NodeName
-	metricLabels[podPhase] = string(pod.Status.Phase)
-	node := &v1.Node{}
-	if err := c.kubeClient.Get(ctx, types.NamespacedName{Name: pod.Spec.NodeName}, node); err != nil {
-		metricLabels[podHostZone] = "N/A"
-		metricLabels[podHostArchitecture] = "N/A"
-		metricLabels[podHostCapacityType] = "N/A"
-		metricLabels[podHostInstanceType] = "N/A"
-		if provisionerName, ok := pod.Spec.NodeSelector[v1alpha5.ProvisionerNameLabelKey]; ok {
-			metricLabels[podProvisioner] = provisionerName
-		} else {
-			metricLabels[podProvisioner] = "N/A"
-		}
-	} else {
-		metricLabels[podHostZone] = node.Labels[v1.LabelTopologyZone]
-		metricLabels[podHostArchitecture] = node.Labels[v1.LabelArchStable]
-		if capacityType, ok := node.Labels[v1alpha5.LabelCapacityType]; !ok {
-			metricLabels[podHostCapacityType] = "N/A"
-		} else {
-			metricLabels[podHostCapacityType] = capacityType
-		}
-		metricLabels[podHostInstanceType] = node.Labels[v1.LabelInstanceTypeStable]
-		if provisionerName, ok := node.Labels[v1alpha5.ProvisionerNameLabelKey]; !ok {
-			metricLabels[podProvisioner] = "N/A"
-		} else {
-			metricLabels[podProvisioner] = provisionerName
-		}
-	}
-	return metricLabels
 }
